fix no EVIDENCE bug in stitch_fragmented_CNVs.sh (#2)
The input VCF lacks the "EVIDENCE" field in the INFO column. As a
result, during the step "Subset VCF to biallelic CNVs without PE or SR
support", the script will produce an empty biallelic_depth_CNVs.vcf.gz
file containing no variants. This step does not raise an error or halt
execution, so the pipeline continues; however, no stitched VCF is
generated. Therefore, the script's output will be identical to the
input.